Initial Commit: a program that allows to code a message into an image, through encoding the ASCII value of each letter of a message into an RGB value of each pixel of a created image.
In other words: if a message is the word "apple":

   a   p    p    l    e
  97  112  112  108  101  //ASCII values

After that an image with 8 pixes will be created (8 in order to achieve a nice looking square) and each pixel will be colored with RGB value of the ASCII letter so in case of letter 'p'; r=1, g=1, b=2; (the 3 unused pixels will remain uninitialised - 0,0,0)

With such low RGB values the image will appear black to a human eye, in case user wants to change that - it is very easy to add 100 to each RGB value in the code.

The programm also supplies a decoder.